fixed bug of Connect function in Socket module

+++ network/select/select-linux/socket.h
@@ -59,7 +59,7 @@ public:
   void Listen(void);
 
   void Accept(Socket* s, struct sockaddr* addr);
-  bool Connect(const char* ip, unsigned int port);
+  bool Connect(const char* ip, unsigned short port);
   int Read(int len, char* buf);
   int Write(const char* buf, int len);
 };
